Replace taxon links with static links
We don't want to link to the taxon finders anymore.  The content that
we now link to is created for this specific purpose.

We've also removed the email signup link as this will now be on the finder
that we're going to link to.
