+++ app/views/brexit_campaign/show.html.erb
@@ -6,20 +6,6 @@
 
 <%= render partial: 'header' %>
 
-<div class="taxon-page__email-link-wrapper">
-  <div class="full-page-width-wrapper">
-      <%= render "components/email-link", {
-        text: t('campaign.email_signup'),
-        href: "https://www.gov.uk/email-signup/?topic=/government/brexit",
-        data: {
-          module: "track-click",
-          track_category: "emailAlertLinkClicked",
-          track_action: "/government/brexit"
-        }
-      } %>
-  </div>
-</div>
-
 <div class="full-page-width-wrapper">
   <%= render partial: 'all_taxons' %>
 </div>
